Lookup imagemagick patch level from sources
Imagemagick only provides the most recent patchlevel for each version.
These changes allow you to specify a version without a patchlevel and
allow Ansible to dynamically determine the pacthlevel from
http://www.imagemagick.org/download/releases/

Additionally, the default archive format is changed to .tar.xz so that
older versions can be installed, since this is the only format that
releases other than the current version are available in.