Remove the --bindir="$HOME/bin" flag from ffmpeg config
AFAICT, this flag causes ffmpeg to install binaries in an
unexpected place. The job says it succeeded, and all supporting
libraries are where they should be, but the actual binaries are
either not being installed at all, or not being installed
where I can find them.

Removing this config option fixes the problem.