Adding nailgun roles for alternative way of running FITS
This alters the fits role to setup the `fits-nailgun.sh` script and
adds an additional nailgun role to setup a `systemd` service for
running fits via nailgun.

The template included in the nailgun role allows you to control
the the fits service like this:

`sudo systemctl service nailgun start`
`sudo systemctl service nailgun stop`

The service will be started when the role is run. This allows
you to run fits like this on the server:

`ng edu.harvard.hul.ois.fits.Fits -i somefile.jpg`