Install ClamAV and FreshClam for virus checking
Install packages necessary to run clamav in daemon mode.
By default Ubuntu runs clamav and freshclam in daemon mode
and freshclam is set by default to download new virus
definitions hourly. It is not necessary (and does not work)
to run freshclam as a separate command.

Remove the RedHat flavors of ClamAV installation since we
can't test those against the rest of ansible-samvera at this time,
but leave them in the history in case we need to support
RedHat at a later date.