Sidekiq as service (#38)
* Run sidekiq as a service

* Rename sudoers file to what the role expects

* Start sidekiq when running role

This commit changes the role so that sidekiq is started
instead of just being enabled. That way it is running when
the role is complete. Enabling it will only start it on a
reboot.

This also changes the role to use the systemd module syntax
instead of running the service commands through shell commands.
For more details on that syntax see:
http://docs.ansible.com/ansible/latest/modules/systemd_module.html