Add fits-servlet to fits role
This adds changes to the fits role so that the
`fits-servlet` application is also installed. This involves
setting up Tomcat to run the war.

The role downloads the war and sets up tomcat
so that it will be run. The last task in the role
checks to see if it is running.

This role can be ran without running the fedora role
which also installs tomcat, or it can be run alongside
that fedora role.