Ensure directory for passenger temp files persists
The default location for passenger temp files (e.g., PID files) differs
between Red Hat and Ubuntu. This causes problems for us because we are
often deploying an application to both environments and it's better if
we can have a standard place.

Our passenger setup is configured to use /var/run/passenger-instreg, but
that directory does not persist across a hard reboot. The recommended
way of fixing that in Ubuntu these days is to add an entry to
/usr/lib/tempdirs.d

This commit adds that file for passenger, which should fix the behavior
we've been seeing that passenger does not restart after reboot
(it's trying to start, but it can't write its PID file so it fails).

References:

http://manpages.ubuntu.com/manpages/xenial/en/man5/tmpfiles.d.5.html
https://serverfault.com/questions/74796/directory-in-var-run-gets-deleted-after-hard-reboot