Fix hostname fqdn bug
**ISSUE**
The command `hostname --fqdn` had been returning only the host
portion of the fully qualified name.  We want access to the fully
qualified name in Rails applicaitons; if we can successfully configure
the system hostname value, we can avoid duplicating static
configuration elsewhere.

**DIAGNOSIS**
The hostname command uses /etc/hosts to determine the friendly name
for 127.0.0.1.  It returns the first entry for that IP address.

**FIX**
Re-order so the fully qualified name comes before the short name in
/etc/hosts